% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woof.R
\name{woof}
\alias{woof}
\title{Structure waldo's output}
\usage{
woof(x)
}
\arguments{
\item{x}{a "waldo_compare" object}
}
\value{
a "woof_compare" object
}
\description{
{woof} is {waldo}'s companion. Given a "waldo_compare" object it returns
a nested list easier to explore and programatically work with. Any subset
of the output prints like \code{waldo:compare()}'s output does.
}
\details{
The output replicates the structure of the original compared objects but :
\itemize{
\item only elements that changed are shown
\item due to the above numeric indexes are converted to character
\item attributes and environments are provided as elements, for easier exploration
}
}
\examples{
foo <- list(
a = "b",
list(c = list("d"), x = "e"),
f = "g"
)

bar <- list(
  a = "B",
  list(c = list("D"), x = "E")
)
x <- waldo_compare(foo, bar)
w <- woof(x)
w$`2`
}
